fix: ensure InvalidNtpServerResponseException propagates properly
If we never got a successful response, and one of the branches in the
flatmap fail and completes, while another parallel one hasn't completed
and emits an error, we are seeing an UndeliverableException.

To simulate this: set a low server delay threshold (make it super low,
thus always failing). you'll notice the app crashes (vs reducing the
number of ntp to just 1 and always failing which properly propagates
the error).

for more context check this:
https://stackoverflow.com/questions/43525052/rxjava2-observable-take-thr
ows-undeliverableexception